Fixed corruption of rotation values for MapObjects
Some MapObjects have a scale of zero on one or more of their axes, which caused problems with storing the rotation values of their transformation matrix when recompiling a dae back into a csb (or even re-saving the dae in another program!)

MapObjects which have a scale of zero will have their scales set to a very small value (1e-9) when exported to .dae.
This value (or any value smaller) will be automatically set back to zero when compiling the .dae back into a .csb.